pyproject.toml
[tool.poetry]
name = "safaribackend"
version = "0.1.0"
description = "A restApi project which follows best practices"
authors = ["Can İlgu <canilguu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
Django = "^4.1.7"
django-debug-toolbar = "^3.8.1"
python-decouple = "^3.7"
django-cleanup = "^7.0.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "0.26.0"
pillow = "^9.4.0"
boto3 = "^1.26.99"
celery = "^5.2.7"
redis = "^4.5.3"
aiohttp = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
